And now with -O2
