+++ plutus-core/plutus-core.cabal
@@ -57,6 +57,8 @@ library
         PlutusCore.Evaluation.Machine.BuiltinCostModel
         PlutusCore.Evaluation.Machine.Ck
         PlutusCore.Evaluation.Machine.CostModelInterface
+        PlutusCore.Evaluation.Machine.CostingFun.Core
+        PlutusCore.Evaluation.Machine.CostingFun.JSON
         PlutusCore.Evaluation.Machine.ExBudget
         PlutusCore.Evaluation.Machine.ExMemory
         PlutusCore.Evaluation.Machine.Exception
