+++ plutus-core/plutus-core/src/PlutusCore/Evaluation/Machine/BuiltinCostModel.hs
@@ -4,12 +4,11 @@
 {-# LANGUAGE DeriveAnyClass       #-}
 {-# LANGUAGE FlexibleInstances    #-}
 {-# LANGUAGE OverloadedStrings    #-}
-{-# LANGUAGE StrictData           #-}
 {-# LANGUAGE TypeFamilies         #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-{-# OPTIONS_GHC -O0           #-}
-
+{-# LANGUAGE StrictData           #-}
+{-# OPTIONS_GHC -O0               #-}
 module PlutusCore.Evaluation.Machine.BuiltinCostModel
     ( BuiltinCostModel
     , BuiltinCostModelBase(..)
@@ -40,16 +39,15 @@ where
 
 import PlutusPrelude hiding (toList)
 
+import PlutusCore.Evaluation.Machine.CostingFun.Core
+import PlutusCore.Evaluation.Machine.CostingFun.JSON ()
 import PlutusCore.Evaluation.Machine.ExBudget
-import PlutusCore.Evaluation.Machine.ExMemory
 
 import Barbies
 import Data.Aeson
 import Data.Default.Class
-import Data.Hashable
 import Data.Kind qualified as Kind
 import Deriving.Aeson
-import GHC.Exts (lazy)
 import Language.Haskell.TH.Syntax hiding (Name, newName)
 
 type BuiltinCostModel = BuiltinCostModelBase CostingFun
@@ -177,397 +175,3 @@ deriving anyclass instance AllArgumentModels Default f => Default (BuiltinCostMo
 deriving stock instance AllArgumentModels Lift    f => Lift    (BuiltinCostModelBase f)
 deriving stock instance AllArgumentModels Show    f => Show    (BuiltinCostModelBase f)
 deriving stock instance AllArgumentModels Eq      f => Eq      (BuiltinCostModelBase f)
-
--- TODO there's probably a nice way to abstract over the number of arguments here. Feel free to implement it.
-
-data CostingFun model = CostingFun
-    { costingFunCpu    :: model
-    , costingFunMemory :: model
-    }
-    deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (Default, NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "costingFun", CamelToSnake)] (CostingFun model)
-
-
----------------- One-argument costing functions ----------------
-
-data ModelOneArgument =
-    ModelOneArgumentConstantCost CostingInteger
-    | ModelOneArgumentLinearCost ModelLinearSize
-    deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[ TagSingleConstructors
-         , SumTaggedObject "type" "arguments"
-         , ConstructorTagModifier (StripPrefix "ModelOneArgument", CamelToSnake)]
-        ModelOneArgument
-        -- Without TagSingleConstructors the format can change unexpectedly if
-        -- you add/remove constructors because you don't get the tags if there's
-        -- only one constructor but you do if there's more than one.
-instance Default ModelOneArgument where
-    def = ModelOneArgumentConstantCost 0
-
-runCostingFunOneArgument
-    :: CostingFun ModelOneArgument
-    -> ExMemory
-    -> ExBudget
-runCostingFunOneArgument (CostingFun cpu mem) = lazy $
-    let !runCpu = runOneArgumentModel cpu
-        {-# NOINLINE runCpu #-}
-        !runMem = runOneArgumentModel mem
-        {-# NOINLINE runMem #-}
-    in \mem1 ->
-        ExBudget (ExCPU    $ runCpu mem1)
-                 (ExMemory $ runMem mem1)
-
-runOneArgumentModel
-    :: ModelOneArgument
-    -> ExMemory
-    -> CostingInteger
-runOneArgumentModel (ModelOneArgumentConstantCost c) = lazy $ \_ -> c
-runOneArgumentModel (ModelOneArgumentLinearCost (ModelLinearSize intercept slope)) = lazy $ \(ExMemory s) ->
-    s * slope + intercept
-
-
----------------- Two-argument costing functions ----------------
-
--- | s * (x + y) + I
-data ModelAddedSizes = ModelAddedSizes
-    { modelAddedSizesIntercept :: CostingInteger
-    , modelAddedSizesSlope     :: CostingInteger
-    } deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "modelAddedSizes", CamelToSnake)] ModelAddedSizes
-
--- | s * (x - y) + I
-data ModelSubtractedSizes = ModelSubtractedSizes
-    { modelSubtractedSizesIntercept :: CostingInteger
-    , modelSubtractedSizesSlope     :: CostingInteger
-    , modelSubtractedSizesMinimum   :: CostingInteger
-    } deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "modelSubtractedSizes", CamelToSnake)] ModelSubtractedSizes
-
-data ModelLinearSize = ModelLinearSize
-    { modelLinearSizeIntercept :: CostingInteger
-    , modelLinearSizeSlope     :: CostingInteger
-    } deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "modelLinearSize", CamelToSnake)] ModelLinearSize
-
--- | s * (x * y) + I
-data ModelMultipliedSizes = ModelMultipliedSizes
-    { modelMultipliedSizesIntercept :: CostingInteger
-    , modelMultipliedSizesSlope     :: CostingInteger
-    } deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "modelMultipliedSizes", CamelToSnake)] ModelMultipliedSizes
-
--- | s * min(x, y) + I
-data ModelMinSize = ModelMinSize
-    { modelMinSizeIntercept :: CostingInteger
-    , modelMinSizeSlope     :: CostingInteger
-    } deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "modelMinSize", CamelToSnake)] ModelMinSize
-
--- | s * max(x, y) + I
-data ModelMaxSize = ModelMaxSize
-    { modelMaxSizeIntercept :: CostingInteger
-    , modelMaxSizeSlope     :: CostingInteger
-    } deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "modelMaxSize", CamelToSnake)] ModelMaxSize
-
--- | if p then s*x else c; p depends on usage
-data ModelConstantOrLinear = ModelConstantOrLinear
-    { modelConstantOrLinearConstant  :: CostingInteger
-    , modelConstantOrLinearIntercept :: CostingInteger
-    , modelConstantOrLinearSlope     :: CostingInteger
-    } deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "modelConstantOrLinear", CamelToSnake)] ModelConstantOrLinear
-
--- | if p then f(x,y) else c; p depends on usage
-data ModelConstantOrTwoArguments = ModelConstantOrTwoArguments
-    { modelConstantOrTwoArgumentsConstant :: CostingInteger
-    , modelConstantOrTwoArgumentsModel    :: ModelTwoArguments
-    } deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[FieldLabelModifier (StripPrefix "modelConstantOrTwoArguments", CamelToSnake)] ModelConstantOrTwoArguments
-
-
-data ModelTwoArguments =
-    ModelTwoArgumentsConstantCost       CostingInteger
-  | ModelTwoArgumentsLinearInX          ModelLinearSize
-  | ModelTwoArgumentsLinearInY          ModelLinearSize
-  | ModelTwoArgumentsAddedSizes         ModelAddedSizes
-  | ModelTwoArgumentsSubtractedSizes    ModelSubtractedSizes
-  | ModelTwoArgumentsMultipliedSizes    ModelMultipliedSizes
-  | ModelTwoArgumentsMinSize            ModelMinSize
-  | ModelTwoArgumentsMaxSize            ModelMaxSize
-  | ModelTwoArgumentsLinearOnDiagonal   ModelConstantOrLinear
-  | ModelTwoArgumentsConstAboveDiagonal ModelConstantOrTwoArguments
-  | ModelTwoArgumentsConstBelowDiagonal ModelConstantOrTwoArguments
-    deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[ TagSingleConstructors
-         , SumTaggedObject "type" "arguments"
-         , ConstructorTagModifier (StripPrefix "ModelTwoArguments", CamelToSnake)]
-        ModelTwoArguments
-
-instance Default ModelTwoArguments where
-    def = ModelTwoArgumentsConstantCost 0
-
-runCostingFunTwoArguments
-    :: CostingFun ModelTwoArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExBudget
-runCostingFunTwoArguments (CostingFun cpu mem) = lazy $
-    let !runCpu = runTwoArgumentModel cpu
-        {-# NOINLINE runCpu #-}
-        !runMem = runTwoArgumentModel mem
-        {-# NOINLINE runMem #-}
-    in \mem1 mem2 ->
-        ExBudget (ExCPU    $ runCpu mem1 mem2)
-                 (ExMemory $ runMem mem1 mem2)
-
-runTwoArgumentModel
-    :: ModelTwoArguments
-    -> ExMemory
-    -> ExMemory
-    -> CostingInteger
-runTwoArgumentModel
-    (ModelTwoArgumentsConstantCost c) = lazy $ \_ _ -> c
-runTwoArgumentModel
-    (ModelTwoArgumentsAddedSizes (ModelAddedSizes intercept slope)) = lazy $ \(ExMemory size1) (ExMemory size2) ->
-        (size1 + size2) * slope + intercept -- TODO is this even correct? If not, adjust the other implementations too.
-runTwoArgumentModel
-    (ModelTwoArgumentsSubtractedSizes (ModelSubtractedSizes intercept slope minSize)) = lazy $ \(ExMemory size1) (ExMemory size2) ->
-        (max minSize (size1 - size2)) * slope + intercept
-runTwoArgumentModel
-    (ModelTwoArgumentsMultipliedSizes (ModelMultipliedSizes intercept slope)) = lazy $ \(ExMemory size1) (ExMemory size2) ->
-        (size1 * size2) * slope + intercept
-runTwoArgumentModel
-    (ModelTwoArgumentsMinSize (ModelMinSize intercept slope)) = lazy $ \(ExMemory size1) (ExMemory size2) ->
-        (min size1 size2) * slope + intercept
-runTwoArgumentModel
-    (ModelTwoArgumentsMaxSize (ModelMaxSize intercept slope)) = lazy $ \(ExMemory size1) (ExMemory size2) ->
-        (max size1 size2) * slope + intercept
-runTwoArgumentModel
-    (ModelTwoArgumentsLinearInX (ModelLinearSize intercept slope)) = lazy $ \(ExMemory size1) (ExMemory _) ->
-        size1 * slope + intercept
-runTwoArgumentModel
-    (ModelTwoArgumentsLinearInY (ModelLinearSize intercept slope)) = lazy $ \(ExMemory _) (ExMemory size2) ->
-        size2 * slope + intercept
-runTwoArgumentModel  -- Off the diagonal, return the constant.  On the diagonal, run the one-variable linear model.
-    (ModelTwoArgumentsLinearOnDiagonal (ModelConstantOrLinear c intercept slope)) = lazy $ \(ExMemory xSize) (ExMemory ySize) ->
-        if xSize == ySize
-        then xSize * slope + intercept
-        else c
-runTwoArgumentModel -- Below the diagonal, return the constant. Above the diagonal, run the other model.
-    (ModelTwoArgumentsConstBelowDiagonal (ModelConstantOrTwoArguments c m)) = lazy $
-        let !run = runTwoArgumentModel m
-            {-# NOINLINE run #-}
-        in \xMem yMem ->
-            if xMem > yMem
-            then c
-            else run xMem yMem
-runTwoArgumentModel -- Above the diagonal, return the constant. Below the diagonal, run the other model.
-    (ModelTwoArgumentsConstAboveDiagonal (ModelConstantOrTwoArguments c m)) = lazy $
-        let !run = runTwoArgumentModel m
-            {-# NOINLINE run #-}
-        in \xMem yMem ->
-            if xMem < yMem
-            then c
-            else run xMem yMem
-
-
----------------- Three-argument costing functions ----------------
-
-data ModelThreeArguments =
-    ModelThreeArgumentsConstantCost CostingInteger
-  | ModelThreeArgumentsAddedSizes   ModelAddedSizes
-  | ModelThreeArgumentsLinearInX    ModelLinearSize
-  | ModelThreeArgumentsLinearInY    ModelLinearSize
-  | ModelThreeArgumentsLinearInZ    ModelLinearSize
-    deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[ TagSingleConstructors
-         , SumTaggedObject "type" "arguments"
-         , ConstructorTagModifier (StripPrefix "ModelThreeArguments", CamelToSnake)]
-        ModelThreeArguments
-
-instance Default ModelThreeArguments where
-    def = ModelThreeArgumentsConstantCost 0
-
-runThreeArgumentModel
-    :: ModelThreeArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> CostingInteger
-runThreeArgumentModel (ModelThreeArgumentsConstantCost c) = lazy $ \_ _ _ -> c
-runThreeArgumentModel (ModelThreeArgumentsAddedSizes (ModelAddedSizes intercept slope)) = lazy $ \(ExMemory size1) (ExMemory size2) (ExMemory size3) -> lazy $
-    (size1 + size2 + size3) * slope + intercept
-runThreeArgumentModel (ModelThreeArgumentsLinearInX (ModelLinearSize intercept slope)) = lazy $ \(ExMemory size1) _ _ -> lazy $
-    size1 * slope + intercept
-runThreeArgumentModel (ModelThreeArgumentsLinearInY (ModelLinearSize intercept slope)) = lazy $ \_ (ExMemory size2) _ -> lazy $
-    size2 * slope + intercept
-runThreeArgumentModel (ModelThreeArgumentsLinearInZ (ModelLinearSize intercept slope)) = lazy $ \_ _ (ExMemory size3) -> lazy $
-    size3 * slope + intercept
-
-runCostingFunThreeArguments
-    :: CostingFun ModelThreeArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExBudget
-runCostingFunThreeArguments (CostingFun cpu mem) = lazy $
-    let !runCpu = runThreeArgumentModel cpu
-        {-# NOINLINE runCpu #-}
-        !runMem = runThreeArgumentModel mem
-        {-# NOINLINE runMem #-}
-    in \mem1 mem2 mem3 ->
-        ExBudget (ExCPU    $ runCpu mem1 mem2 mem3)
-                 (ExMemory $ runMem mem1 mem2 mem3)
-
-
----------------- Four-argument costing functions ----------------
-
-data ModelFourArguments =
-      ModelFourArgumentsConstantCost CostingInteger
-    deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[ TagSingleConstructors
-         , SumTaggedObject "type" "arguments"
-         , ConstructorTagModifier (StripPrefix "ModelFourArguments", CamelToSnake)]
-        ModelFourArguments
-
-instance Default ModelFourArguments where
-    def = ModelFourArgumentsConstantCost 0
-
-runFourArgumentModel
-    :: ModelFourArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> CostingInteger
-runFourArgumentModel (ModelFourArgumentsConstantCost c) = lazy $ \_ _ _ _ -> c
-
-runCostingFunFourArguments
-    :: CostingFun ModelFourArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExBudget
-runCostingFunFourArguments (CostingFun cpu mem) = lazy $
-    let !runCpu = runFourArgumentModel cpu
-        {-# NOINLINE runCpu #-}
-        !runMem = runFourArgumentModel mem
-        {-# NOINLINE runMem #-}
-    in \mem1 mem2 mem3 mem4 ->
-        ExBudget (ExCPU    $ runCpu mem1 mem2 mem3 mem4)
-                 (ExMemory $ runMem mem1 mem2 mem3 mem4)
-
-
----------------- Five-argument costing functions ----------------
-
-data ModelFiveArguments =
-      ModelFiveArgumentsConstantCost CostingInteger
-    deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[ TagSingleConstructors
-         , SumTaggedObject "type" "arguments"
-         , ConstructorTagModifier (StripPrefix "ModelFiveArguments", CamelToSnake)]
-        ModelFiveArguments
-
-instance Default ModelFiveArguments where
-    def = ModelFiveArgumentsConstantCost 0
-
-runFiveArgumentModel
-    :: ModelFiveArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> CostingInteger
-runFiveArgumentModel (ModelFiveArgumentsConstantCost c) = lazy $ \_ _ _ _ _ -> c
-
-runCostingFunFiveArguments
-    :: CostingFun ModelFiveArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExBudget
-runCostingFunFiveArguments (CostingFun cpu mem) = lazy $
-    let !runCpu = runFiveArgumentModel cpu
-        {-# NOINLINE runCpu #-}
-        !runMem = runFiveArgumentModel mem
-        {-# NOINLINE runMem #-}
-    in \mem1 mem2 mem3 mem4 mem5 ->
-        ExBudget (ExCPU    $ runCpu mem1 mem2 mem3 mem4 mem5)
-                 (ExMemory $ runMem mem1 mem2 mem3 mem4 mem5)
-
----------------- Six-argument costing functions ----------------
-
-data ModelSixArguments =
-      ModelSixArgumentsConstantCost CostingInteger
-    deriving stock (Show, Eq, Generic, Lift)
-    deriving anyclass (NFData)
-    deriving (FromJSON, ToJSON) via CustomJSON
-        '[ TagSingleConstructors
-         , SumTaggedObject "type" "arguments"
-         , ConstructorTagModifier (StripPrefix "ModelSixArguments", CamelToSnake)]
-        ModelSixArguments
-
-instance Default ModelSixArguments where
-    def = ModelSixArgumentsConstantCost 0
-
-runSixArgumentModel
-    :: ModelSixArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> CostingInteger
-runSixArgumentModel (ModelSixArgumentsConstantCost c) = lazy $ \_ _ _ _ _ _ -> c
-
-runCostingFunSixArguments
-    :: CostingFun ModelSixArguments
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExMemory
-    -> ExBudget
-runCostingFunSixArguments (CostingFun cpu mem) = lazy $
-    let !runCpu = runSixArgumentModel cpu
-        {-# NOINLINE runCpu #-}
-        !runMem = runSixArgumentModel mem
-        {-# NOINLINE runMem #-}
-    in \mem1 mem2 mem3 mem4 mem5 mem6 ->
-        ExBudget (ExCPU    $ runCpu mem1 mem2 mem3 mem4 mem5 mem6)
-                 (ExMemory $ runMem mem1 mem2 mem3 mem4 mem5 mem6)
